// frontend/src/api/generated/models/ShareListItem.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ShareListItem = {
    available: boolean;
    created_at: string;
    created_by: string;
    creator_name: string;
    download_count: number;
    expires_at: (string | null);
    id: string;
    max_downloads: (number | null);
    max_views: (number | null);
    permission: string;
    root_name: string;
    root_node_id: string;
    root_node_type: string;
    share_type: string;
    status: string;
    view_count: number;
};